services/data_upload/requirements.txt
fastapi
uvicorn
pandas
azure-storage-blob
python-json-logger
requests
httpx
